requirements.txt
pycocotools
numpy
scikit-image
matplotlib
pylab-sdk
requests
opencv-python